Realised there is a problem in the DB Structure. Fix by adding treeVersion to NodeOption.
Before the public ID on Node Option was per Node.

Realised this is a pain as there are then places you need to know the node option public ID and the node it came from. It would be better if we could just have the Public ID.

So made the public ID on Node Option per Tree Version, the same as the public ID on Node.

Existing data will have to be checked by hand to make sure there are no clashes.